+++ src/main/groovy/groovyx/groovyserv/ExitStatus.groovy
@@ -21,16 +21,15 @@ package groovyx.groovyserv
 public enum ExitStatus {
 
     SUCCESS(0),
-    INTERRUPTED(1),
+    UNEXPECTED_ERROR(1),
     INVALID_REQUEST(2),
     IO_ERROR(3),
     ILLEGAL_STATE(4),
     TERMINATED(5),
-    COMMAND_ERROR(6),
-    INVALID_AUTHTOKEN(201),
-    CLIENT_NOT_ALLOWED(202),
-    UNEXPECTED_ERROR(-1),
-    FORCELY_SHUTDOWN(99)
+    INTERRUPTED(6),
+    FORCELY_SHUTDOWN(7),
+    INVALID_AUTHTOKEN(201), // NOTE: no change because it's used from a user command
+    CLIENT_NOT_ALLOWED(202) // NOTE: no change because it's used from a user command
 
     int code