+++ src/main/java/groovyx/groovyserv/GroovyMain2.java
@@ -128,7 +128,7 @@ public class GroovyMain2 {
                 // If we fail, then exit with an error so scripting frameworks can catch it
                 // TODO: pass printstream(s) down through process
                 if (!process(cmd, classpath)) {
-                    //System.exit(1); // for GroovyServ: disabled because this causes a secondary disaster
+                    System.exit(1);
                 }
             }
         } catch (ParseException pe) {